Update README.md
In the current meta, people with good memories are often seen as knowledgeable, smart, competent, and dependable. 
Memory Trigger - A DosBox (Turbo C++) based desktop game app is aimed at helping improve their remembering power in a fun and easy way.
The player is shown a randomized sequence on a grid consisting of 6 tiles, the players need to remember and repeat the sequences shown
The game is for both students as well as employees to improve their concentration level.